osd: uint64/size_t casts for p2roundup/max/min on macOS
ECUtil.h (p2roundup with EC_ALIGN_SIZE), ECExtentCache.h (std::max MIN_LINE_SIZE
vs get_chunk_size), ReplicatedBackend.cc (std::min available/num_new_bytes):
same size_t-vs-uint64_t distinct-type issue on Darwin. The two EC headers are
included by most osd TUs, so this clears the bulk of the osd-layer errors.

+++ src/osd/ECUtil.h
@@ -99,7 +99,7 @@ class slice_iterator {
       uint64_t off = 0;
       char *c_str = bp.c_str();
       // Skip any non-aligned chunk.
-      uint64_t analysed = p2roundup((uintptr_t)c_str, EC_ALIGN_SIZE) - (uintptr_t)c_str;
+      uint64_t analysed = p2roundup((uintptr_t)c_str, (uintptr_t)EC_ALIGN_SIZE) - (uintptr_t)c_str;
 
       while (off + analysed <= bp_len) {
         bool new_is_zeros;
